LevelLoader Update
It has some warnings but should work for now. In order to add a scene, instead of build index, you will have to provide a name, as well as a spawn location for the player in the next scene.

There is a bug at the moment where the player is invisible for a few seconds but I think it should be able to be worked out eventually, and isn't game breaking atm.